+++ src/dataprotection/azext_dataprotection/manual/commands.py
@@ -7,6 +7,7 @@
 # pylint: disable=too-many-locals
 # pylint: disable=line-too-long
 
+from azure.cli.command_modules.storage._client_factory import cf_blob_container_mgmt
 from azext_dataprotection.manual._client_factory import cf_resource_graph_client
 from ._exception_handler import exception_handler
 
@@ -15,12 +16,14 @@ def load_command_table(self, _):
 
     # Manual backup-instance commmands
     with self.command_group('dataprotection backup-instance', exception_handler=exception_handler) as g:
-        g.custom_command('initialize-backupconfig', "dataprotection_backup_instance_initialize_backupconfig")
         g.custom_command('initialize-restoreconfig', "dataprotection_backup_instance_initialize_restoreconfig")
         g.custom_command('initialize', "dataprotection_backup_instance_initialize")
         g.custom_command('list-from-resourcegraph', 'dataprotection_backup_instance_list_from_resourcegraph', client_factory=cf_resource_graph_client)
         g.custom_command('update-msi-permissions', 'dataprotection_backup_instance_update_msi_permissions')
         g.custom_command('update-policy', "dataprotection_backup_instance_update_policy", supports_no_wait=True)
+
+    with self.command_group('dataprotection backup-instance', exception_handler=exception_handler, client_factory=cf_blob_container_mgmt) as g:
+        g.custom_command('initialize-backupconfig', "dataprotection_backup_instance_initialize_backupconfig")
 
     # Migrated to AAZ, using a function based override approach. Function-based override uses AAZ arg schema to validate input
     # but is less manageable. Needs manual management of params and help. Find a way to use class-based override without compromising